+++ volmdlr/faces.py
@@ -5991,6 +5991,7 @@ class BSplineSurface3D(Surface3D):
 
         :param other_bspline_surface3d: Other adjacent surface
         :type other_bspline_surface3d: :class:`volmdlr.faces.BSplineSurface3D`
+
         :return: Merged surface
         :rtype: :class:`volmdlr.faces.BSplineSurface3D`
         """
@@ -6005,20 +6006,17 @@ class BSplineSurface3D(Surface3D):
                   other_bspline_face3d.surface2d.outer_contour.center_of_mass()]
         grid2d_direction = (bspline_face3d.pair_with(other_bspline_face3d))[1]
 
-        # if bspline_face3d.outer_contour3d.is_sharing_primitives_with(other_bspline_face3d.outer_contour3d):
-        #
-        #     xmin, xmax, ymin, ymax = self.xy_limits(other_bspline_surface3d)
-
-        if self.is_intersected_with(other_bspline_surface3d):
-            # find primitives to split with
+        if (not bspline_face3d.outer_contour3d.is_sharing_primitives_with(other_bspline_face3d.outer_contour3d)
+            and self.is_intersected_with(other_bspline_surface3d)):
+            # find pimitives to split with
             contour1 = bspline_face3d.outer_contour3d
             contour2 = other_bspline_face3d.outer_contour3d
 
             distances = []
-            for prim_1 in contour1.primitives:
+            for prim1 in contour1.primitives:
                 dis = []
                 for prim2 in contour2.primitives:
-                    point1 = (prim_1.start + prim_1.end) / 2
+                    point1 = (prim1.start + prim1.end) / 2
                     point2 = (prim2.start + prim2.end) / 2
                     dis.append(point1.point_distance(point2))
                 distances.append(dis)
@@ -6038,33 +6036,31 @@ class BSplineSurface3D(Surface3D):
 
                 bsplines_new[i] = surfaces[errors.index(min(errors))]
 
-            # xmin, xmax, ymin, ymax = [0] * len(bsplines_new), [1] * len(bsplines_new), [0] * \
-            #     len(bsplines_new), [1] * len(bsplines_new)
-
             grid2d_direction = (
                 bsplines_new[0].rectangular_cut(
                     0, 1, 0, 1).pair_with(
                     bsplines_new[1].rectangular_cut(
                         0, 1, 0, 1)))[1]
 
-        # else:
-        #     xmin, xmax, ymin, ymax = [0] * len(bsplines_new), [1] * len(bsplines_new), [0] * \
-        #                              len(bsplines_new), [1] * len(bsplines_new)
-
         # grid3d
+        nb = 10
         points3d = []
+        is_true = (bspline_face3d.outer_contour3d.is_sharing_primitives_with(other_bspline_face3d.outer_contour3d)
+                    or self.is_intersected_with(other_bspline_surface3d))
+
         for i, bspline in enumerate(bsplines_new):
             grid3d = bspline.grid3d(volmdlr.grid.Grid2D.from_properties(x_limits=(0, 1),
                                                                         y_limits=(0, 1),
-                                                                        points_nbr=(50, 50),
+                                                                        points_nbr=(nb, nb),
                                                                         direction=grid2d_direction[i]))
 
-            points3d.extend(grid3d)
+            if is_true and i == 1:
+                points3d.extend(grid3d[nb:nb * nb])
+            else:
+                points3d.extend(grid3d)
 
-            # fitting
-        size_u, size_v, degree_u, degree_v = 100, 50, max(
-            bsplines[0].degree_u, bsplines[1].degree_u), max(
-            bsplines[0].degree_v, bsplines[1].degree_v)
+        # fitting
+        size_u, size_v, degree_u, degree_v = (nb * 2) - 1, nb, 3, 3
 
         merged_surface = BSplineSurface3D.points_fitting_into_bspline_surface(
             points3d, size_u, size_v, degree_u, degree_v)
@@ -7284,8 +7280,12 @@ class PlaneFace3D(Face3D):
         min_distance = math.inf
         for edge1 in self.outer_contour3d.primitives:
             for edge2 in other_plane_face.outer_contour3d.primitives:
-                dist = edge1.minimum_distance(edge2,
-                                              return_points=return_points)
+                if hasattr(edge1, 'minimum_distance'):
+                    dist = edge1.minimum_distance(edge2, return_points=return_points)
+                elif hasattr(edge2, 'minimum_distance'):
+                    dist = edge2.minimum_distance(edge1, return_points=return_points)
+                else:
+                    raise AttributeError(f'Neither {edge1} nor {edge2} has a minimum_distance method.')
                 if return_points:
                     if dist[0] < min_distance:
                         min_distance = dist[0]
@@ -8832,8 +8832,11 @@ class BSplineFace3D(Face3D):
             grid2d_direction = [['-y', '-x'], ['-y', '+x']]
 
         elif u1 == 1 and u2 == 1:
-            corresponding_directions.append(('+u', '-v'))
-            grid2d_direction = [['+y', '+x'], ['+y', '-x']]
+            if corresponding_directions == [('+v', '-v')]:
+                grid2d_direction = [['+y', '+x'], ['-y', '-x']]
+            else:
+                grid2d_direction = [['+y', '+x'], ['+y', '-x']]
+            corresponding_directions.append(('+u', '-u'))
 
         elif u1 == 0 and u2 == 1:
             corresponding_directions.append(('+u', '+u'))